Implement tailoring ID and XML namesapce
The ID, as mentioned by the standard should be in the following form:

    xccdf_<namespace>_<object>_<name>

So, for profiles it is:

    xccdf_<namespace>_profile_<name>

and for tailorings it is:

    xccdf_<namespace>_tailoring_<name>

With this in mind, we defined the namespace to be
`compliance.openshift.io` and so this is what's persisted in the
resulting tailoring file for both the Profile and the Tailoring object.

On the other hand, the required XML namespace was added to the Tailoring
object.