Implement support for tailoring variables
Implements a new CR called Variable that implements a subset of the
XCCDF standard related to setting variables during tailoring and
rendering them to a tailoring file using the "set-value" tailoring
element. The subset supported roughly equals to variable types and a
selection of allowed values.

The variables are parsed along with the type and selections in the
profileparser pod. Note that the variable parsing is implemented in a
new profileparser package that is just called from the main package of
the profileparser pod. A future patch should move the rest of the main
package there also to make it testable.

See
`deploy/crds/compliance.openshift.io_v1alpha1_tailoredprofile_cr.yaml`
for an example on how to tune a variable. The tailoredprofile controller
looks up the referenced variables and creates a slice of new Variable
instances passed to the tailoring engine which finally renders the
tailoring XML.

Note that the validation of variable types and selections is done in the
tailoredprofile when looking up the variable and matching it with the
TP's spec: this is done so that we can surface the errors to the TP
status and at the same time have the full information in the Variable CR
available in tailoring if needed.

Closes: #3