TailoredProfiles: Add outputType and remove hardcoded ConfigMap references
TailoredProfiles now take an `outputType` parameter that indicates what
sort of object should be generated from them. Currently only ConfigMaps
are supported, but the intention is to add policies here too.

This also removes the configMap reference from the status, using a more
generic `outputRef` which will contain the name reference to the object that
was created.